use net.ErrClosed (for Go 1.16)
