+++ internal/qerr/errors.go
@@ -29,11 +29,6 @@ func NewCryptoError(tlsAlert uint8, errorMessage string) *TransportError {
 	}
 }
 
-func (e *TransportError) Is(target error) bool {
-	_, ok := target.(*TransportError)
-	return ok
-}
-
 func (e *TransportError) Error() string {
 	str := e.ErrorCode.String()
 	if e.FrameType != 0 {
@@ -63,11 +58,6 @@ type ApplicationError struct {
 
 var _ error = &ApplicationError{}
 
-func (e *ApplicationError) Is(target error) bool {
-	_, ok := target.(*ApplicationError)
-	return ok
-}
-
 func (e *ApplicationError) Error() string {
 	if len(e.ErrorMessage) == 0 {
 		return fmt.Sprintf("Application error %#x", e.ErrorCode)
@@ -82,10 +72,6 @@ var _ error = &IdleTimeoutError{}
 func (e *IdleTimeoutError) Timeout() bool   { return true }
 func (e *IdleTimeoutError) Temporary() bool { return false }
 func (e *IdleTimeoutError) Error() string   { return "timeout: no recent network activity" }
-func (e *IdleTimeoutError) Is(target error) bool {
-	_, ok := target.(*IdleTimeoutError)
-	return ok
-}
 
 type HandshakeTimeoutError struct{}
 
@@ -94,10 +80,6 @@ var _ error = &HandshakeTimeoutError{}
 func (e *HandshakeTimeoutError) Timeout() bool   { return true }
 func (e *HandshakeTimeoutError) Temporary() bool { return false }
 func (e *HandshakeTimeoutError) Error() string   { return "timeout: handshake did not complete in time" }
-func (e *HandshakeTimeoutError) Is(target error) bool {
-	_, ok := target.(*HandshakeTimeoutError)
-	return ok
-}
 
 // A VersionNegotiationError occurs when the client and the server can't agree on a QUIC version.
 type VersionNegotiationError struct {
@@ -109,11 +91,6 @@ func (e *VersionNegotiationError) Error() string {
 	return fmt.Sprintf("no compatible QUIC version found (we support %s, server offered %s)", e.Ours, e.Theirs)
 }
 
-func (e *VersionNegotiationError) Is(target error) bool {
-	_, ok := target.(*VersionNegotiationError)
-	return ok
-}
-
 // A StatelessResetError occurs when we receive a stateless reset.
 type StatelessResetError struct {
 	Token protocol.StatelessResetToken
@@ -123,11 +100,6 @@ var _ net.Error = &StatelessResetError{}
 
 func (e *StatelessResetError) Error() string {
 	return fmt.Sprintf("received a stateless reset with token %x", e.Token)
-}
-
-func (e *StatelessResetError) Is(target error) bool {
-	_, ok := target.(*StatelessResetError)
-	return ok
 }
 
 func (e *StatelessResetError) Timeout() bool   { return false }
